+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Breaking Changes
 
 -   **terra-stopwatch** has been removed.
+-   **tc-checkbox-group** has been removed.
 
 # 11.6.0 (12.07.2021)