+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 -   **terra-stopwatch** has been removed.
 -   **terra-indicator** has been removed.
+-   **tc-checkbox-group** has been removed.
 -   **terra-slider** has been removed.
 
 # 11.6.0 (12.07.2021)