+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     -   updated all angular packages to version 12
     -   update `angular-l10n` to version 12
     -   update `ngx-quill` to version 14
+    -   update `ngx-bootstrap` to version 6.2.0
 
 # 12.0.0-next.0 (13.07.2021)