+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     -   update `angular-l10n` to version 12
     -   update `ngx-quill` to version 14
     -   update `ngx-bootstrap` to version 6.2.0
+    -   update `ckeditor4-angular` to version 2
 
 # 12.0.0-next.0 (13.07.2021)