+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Breaking Changes
 
 -   **terra-stopwatch** has been removed.
+-   **terra-indicator** has been removed.
 -   **tc-checkbox-group** has been removed.
 -   **terra-slider** has been removed.