+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Breaking Changes
 
 -   **terra-stopwatch** has been removed.
+-   **terra-indicator** has been removed.
 -   **tc-checkbox-group** has been removed.
 -   **terra-slider** has been removed.
 -   **interactables** `TerraInteractModule`, `TerraDraggableDirective` and all related interfaces have been removed. See https://github.com/plentymarkets/terra-components/tree/v11.6.0/src/lib/components/interactables for reference.